feat: Implement half-complete files CRUD

+++ salgu-back-end/src/files/files.controller.ts
@@ -7,6 +7,8 @@ import {
   Param,
   Delete,
   UseInterceptors,
+  UseGuards,
+  UnauthorizedException,
 } from '@nestjs/common';
 import { FilesService } from './files.service';
 import { CreateFileDto } from './dto/create-file.dto';
@@ -15,34 +17,60 @@ import { FileEntity } from './entities/file.entity';
 import { FileInterceptor } from '@nestjs/platform-express';
 import { UploadedFile } from '@nestjs/common';
 import { ApiTags } from '@nestjs/swagger';
+import { JwtAuthGuard } from 'src/auth/jwt-auth.guard';
+import { User } from 'src/users/user.decorator';
+import { UserEntity } from 'src/users/entities/user.entity';
 
 @ApiTags('files')
 @Controller('files')
+@UseGuards(JwtAuthGuard)
 export class FilesController {
   constructor(private readonly filesService: FilesService) {}
 
   @Post()
-  create(@Body() createFileDto: CreateFileDto): Promise<FileEntity> {
-    return this.filesService.create(createFileDto);
+  create(
+    @Body() createFileDto: CreateFileDto,
+    @User() user: UserEntity,
+  ): Promise<FileEntity> {
+    return this.filesService.create(createFileDto, user.id);
   }
 
   @Get()
-  findAll() {
-    return this.filesService.findAll();
+  findAll(@User() user: UserEntity) {
+    return this.filesService.findAll(user.id);
   }
 
   @Get(':id')
-  findOne(@Param('id') id: string) {
-    return this.filesService.findOne(+id);
+  async findOne(@Param('id') id: string, @User() user: UserEntity) {
+    const file = await this.filesService.findOne(+id);
+
+    if (file?.ownerId !== user.id || !user.isAdmin)
+      return new UnauthorizedException();
+
+    return file;
   }
 
   @Patch(':id')
-  update(@Param('id') id: string, @Body() updateFileDto: UpdateFileDto) {
+  async update(
+    @Param('id') id: string,
+    @Body() updateFileDto: UpdateFileDto,
+    @User() user: UserEntity,
+  ) {
+    const file = await this.filesService.findOne(+id);
+
+    if (file?.ownerId !== user.id || !user.isAdmin)
+      return new UnauthorizedException();
+
     return this.filesService.update(+id, updateFileDto);
   }
 
   @Delete(':id')
-  remove(@Param('id') id: string) {
+  async remove(@Param('id') id: string, @User() user: UserEntity) {
+    const file = await this.filesService.findOne(+id);
+
+    if (file?.ownerId !== user.id || !user.isAdmin)
+      return new UnauthorizedException();
+
     return this.filesService.remove(+id);
   }
